# By Hal Stroemeria
# Via Hal Stroemeria
* 'Hal' of https://github.com/hal1437/Procon26: (33 commits)
  [Refuctoring]Refucted GetProblem function
  [Added]Added GetProblem from server
  [Refuctoring]Refuctoring main.cpp.
  [Optimize]Optimized heuristic-function
  [Added]Added AntiDensityAround heuristic-function.
  [Added]Changed MultiThread.
  [Added] Added Random heuristic-function
  [Added]Added DoubleLimit solver-class
  [Added]Added Frame heuristic-function.
  [Added]Added QuickBeam solver class
  [Added] BestBeam perfect-function changed to use log.
  [Changed] Changed SD heuristic-function.check to filled cell.
  [Added]Connected perfect-function to BestBeam.
  [Changed] Heuristic-function change to return floating point.
  [Added]Added SD heuristic-function in BestBeam.
  [Added]Changed compare-function.
  [Changed]BestBeam changed to "wide".
  [BugFix]Move to TimeWait.h
  [Added]Added BestBeam solve class
  [Added]Added PPPP Perfect function
  ...

Conflicts:
	Answer.txt
	Hal/Answer.txt
	Hal/CMakeCache.txt
	Hal/CMakeFiles/CMakeDirectoryInformation.cmake
	Hal/CMakeFiles/CMakeOutput.log
	Hal/CMakeFiles/Makefile.cmake
	Hal/CMakeFiles/Makefile2
	Hal/CMakeFiles/Procon26-Hal.dir/CXX.includecache
	Hal/CMakeFiles/Procon26-Hal.dir/DependInfo.cmake
	Hal/CMakeFiles/Procon26-Hal.dir/Solver/BestFirst.cpp.o
	Hal/CMakeFiles/Procon26-Hal.dir/Solver/DLS.cpp.o
	Hal/CMakeFiles/Procon26-Hal.dir/Users/hal1437/Programs/Procon26/Structure/Answer.cpp.o
	Hal/CMakeFiles/Procon26-Hal.dir/Users/hal1437/Programs/Procon26/Structure/Point.cpp.o
	Hal/CMakeFiles/Procon26-Hal.dir/Users/hal1437/Programs/Procon26/Structure/Transform.cpp.o
	Hal/CMakeFiles/Procon26-Hal.dir/build.make
	Hal/CMakeFiles/Procon26-Hal.dir/cmake_clean.cmake
	Hal/CMakeFiles/Procon26-Hal.dir/depend.internal
	Hal/CMakeFiles/Procon26-Hal.dir/depend.make
	Hal/CMakeFiles/Procon26-Hal.dir/flags.make
	Hal/CMakeFiles/Procon26-Hal.dir/link.txt
	Hal/CMakeFiles/Procon26-Hal.dir/progress.make
	Hal/CMakeFiles/progress.marks
	Hal/cmake_install.cmake
	Hal/main.cpp
	Hal/test.txt

+++ Hal/Heuristics/Density.cpp
@@ -1,7 +1,7 @@
 
 #include "Density.h"
 
-unsigned Density::Execution(const Field& field){
+double Density::Execution(const Field& field){
 	int score = 0;
 	for(int i=0;i<FIELD_HEIGHT;i++){
 		for(int j=0;j<FIELD_WIDTH;j++){
